Add handling for in-band DCS tmux:set control messages
During the server's input stream processor, we intercept ANSI DCS
messages beginning with the magic string "tmux:set;". Immediately
following this magic, we expect a string of the form "a=b", followed by
the ST sequence (ESC \). Upon parsing "a=b", we set a global session
variable "@dcs_a" to hold the value "b".

The variable is uniformly prefixed with "@dcs_" to prevent any software
not under the user's control from maliciously controlling tmux in an
unpremeditated way.

An example of setting "@dcs_foo" to the value "bar":
	sh$ printf '\ePtmux:set;foo=bar\e\\'

This mechanism allows tmux to receive state information from local or
remote processes, in band, without assuming the ability to run tmux on
the local tmux socket.  You could use this, for example, to receive
current working context or exception state from a remote system to which
you're connected via ssh, and display it in your status bar.